rename: table컴포넌트 스타일 폴더명 수정 (#43)
