+++ shared/ui/table/statistics/styles.module.scss
