+++ shared/ui/table/vertical/styles.module.scss
